+++ src/app-server.cjs
@@ -110,6 +110,7 @@ function createReviewbotServer(options = {}) {
         loadUsageEvents: options.loadUsageEvents,
         loadBudgetPolicies: options.loadBudgetPolicies,
         loadJobEvents: options.loadJobEvents,
+        loadRunClaims: options.loadRunClaims,
         loadAdminStatus: options.loadAdminStatus,
         authorizeUsageApiAdmin: options.authorizeUsageApiAdmin,
         logger,
@@ -154,6 +155,7 @@ async function handleHttpRequest(request, options) {
         loadUsageEvents: options.loadUsageEvents,
         loadBudgetPolicies: options.loadBudgetPolicies,
         loadJobEvents: options.loadJobEvents,
+        loadRunClaims: options.loadRunClaims,
         loadAdminStatus: options.loadAdminStatus,
         authorizeAdmin: options.authorizeUsageApiAdmin,
       }